Refactors and adds the basic files.
Each kind of function is now in its own file,
and each kind has an example.

The plot.lisp file is responsible for the points
building, i.e. it builds the necessary points
for each plot.

The xml-reader.lisp file is responsible for returning
whatever data is necessary from the parsed XML.

The gnuplot.lisp file is responsible for drawing in
the final file.